pyproject.toml
[project]
name = "lotto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "pytz>=2025.2",
    "scrapy>=2.13.3",
]
